refactor: Renamed the utilities folder
