+++ src/utils/components.ts
@@ -1,0 +1,2 @@
+export * from './components/get-component-file-path.js';
+export * from './components/transform-entity-name.js';
