--- src/utils/files.ts
@@ -1,2 +1,0 @@
-export * from './files/get-component-file-path.js';
-export * from './files/transform-entity-name.js';
